requirements.txt
fastapi[standard]
uvicorn
requests
python-multipart
google-auth
google-api-python-client